refactor: 로또 머신 컨트롤러 생성
